Get lfrestart to work with -X restart
git-svn-id: svn://irodssvn.ucsd.edu/trunk@3996 ea137bd6-aa86-47b6-b90d-b101d9f19a90

+++ iRODS/lib/core/src/rcMisc.c
@@ -2554,7 +2554,10 @@ int deleteFlag)
 	        clearKeyVal (&dataObjInp.condInput);
 	    }
 	} else if (objType == LOCAL_FILE_T) {
-	    status = unlink (restartPath);
+	    if (conn->fileRestart.info.status != FILE_RESTARTED ||
+                strcmp (conn->fileRestart.info.fileName, restartPath) != 0) {
+	        status = unlink (restartPath);
+	    }
 	} else {
             rodsLog (LOG_ERROR,
 	      "setStateForResume: illegal objType %d for %s",
